Handle file reuploads.

+++ src/main/java/org/magneato/resources/UploadResource.java
@@ -1,32 +1,11 @@
 package org.magneato.resources;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.fasterxml.jackson.databind.node.ValueNode;
 import io.dropwizard.views.View;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.FileSystems;
-import java.nio.file.Files;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-
-import javax.annotation.security.RolesAllowed;
-import javax.servlet.http.HttpServletRequest;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.SecurityContext;
-
 import org.apache.commons.io.FilenameUtils;
 import org.glassfish.jersey.media.multipart.FormDataBodyPart;
 import org.glassfish.jersey.media.multipart.FormDataParam;
@@ -39,11 +18,18 @@ import org.magneato.utils.UploadHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.fasterxml.jackson.databind.node.ValueNode;
+import javax.annotation.security.RolesAllowed;
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.*;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.SecurityContext;
+import java.io.*;
+import java.nio.file.FileSystems;
+import java.nio.file.Files;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
 
 // https://github.com/wdawson/dropwizard-auth-example/blob/master/pom.xml
 @Path("/")
@@ -52,12 +38,13 @@ public class UploadResource {
     private static final int SUBDIR_SIZE = 3; // TODO user define
     private static final String GPXTYPE = "application/octet-stream";
     private final ObjectMapper mapper = new ObjectMapper();
-    
+
     PageUtils pageUtils;
 
     private final Logger log = LoggerFactory.getLogger(this.getClass()
             .getName());
 
+    // add to config
     public final static String IMAGEPATH = "/library/images";
     private String imageDir = null;
 
@@ -131,7 +118,7 @@ public class UploadResource {
 
         // store images in a subdir based on up to the first x letters of the
         // filename, avoids putting too many files in one directory... maybe
-        // there is a better idea?
+        // there is a better idea? TODO: change this random string
 
         // TODO make configurable for big sites
         String subDir = FilenameUtils.getBaseName(fileName);
@@ -141,23 +128,25 @@ public class UploadResource {
 
         String name = imageDir + subDir + fileName;
         java.nio.file.Path outputPath = FileSystems.getDefault().getPath(name);
-
-        // create a thumbnail
         long len = 0;
         try {
-            // make the directory, if it doesn't exist
-            Files.createDirectories(outputPath.getParent());
-
-            len = Files.copy(fileInputStream, outputPath);
-            System.out.println("file len " + len);
-            thumbName = UploadHandler.createThumbnail(imageDir + subDir,
-                    fileName, mimeType);
-            System.out.println("thumb " + thumbName);
-            UploadHandler.getMetaData(outputPath.toString());
+            if (Files.exists(outputPath)) {
+                File file = outputPath.toFile();
+                len = file.length();
+                thumbName = UploadHandler.createThumbnail(imageDir + subDir,
+                        fileName, mimeType, true);
+            } else {
+                // make the directory, if it doesn't exist
+                Files.createDirectories(outputPath.getParent());
+                len = Files.copy(fileInputStream, outputPath);
+                System.out.println("file len " + len);
+                thumbName = UploadHandler.createThumbnail(imageDir + subDir,
+                        fileName, mimeType, false);
+                System.out.println("thumb " + thumbName);
+                UploadHandler.getMetaData(outputPath.toString());
+            }
         } catch (IOException e) {
-            log.warn("cannot upload, file already exists " + e.getMessage());
-            // TODO: do something with errors
-            // need to ignore but need thumbName and len anyway
+            log.warn("problem uploading  file " + e.getMessage());
         }
 
         String url = IMAGEPATH + "/" + subDir + fileName;
@@ -231,29 +220,28 @@ public class UploadResource {
                     (int) gpxParser.getMinElevation(),
                     (int) gpxParser.getMaxElevation(), gpxParser.getStartLat(),
                     gpxParser.getStartLon(), uploadInfo.toJson(),
-                    metaData.toJson(),gpxParser.getDate());
+                    metaData.toJson(), gpxParser.getDate());
 
             System.out.println("json " + content + " parent " + parent);
             parent = "ra0a7f77cbee7";
             if (parent != null && !parent.isEmpty()) {
                 String parentJSON = repository.get(parent);
                 log.debug("parent " + parentJSON);
-				try {
-					JsonNode root = mapper.readTree(parentJSON);
-					JsonNode contentTree = mapper.readTree(content);
-					HashMap<String, String> clonableItems = addKeys("", root, false);
-	    			for (Map.Entry<String, String> entry : clonableItems.entrySet()) {
-	    				System.out.println(entry.getKey() + " " + entry.getValue());
-	    				setTokenValue(contentTree, entry.getKey(), entry.getValue());
-	    			}//
-	    			
-	    			content = contentTree.toString();
-				} catch (IOException e) {
-					log.error("Couldn't read parent JSON {1}", e.getMessage());
-					// just ignore, we've got gpx data
+                try {
+                    JsonNode root = mapper.readTree(parentJSON);
+                    JsonNode contentTree = mapper.readTree(content);
+                    HashMap<String, String> clonableItems = addKeys("", root, false);
+                    for (Map.Entry<String, String> entry : clonableItems.entrySet()) {
+                        System.out.println(entry.getKey() + " " + entry.getValue());
+                        setTokenValue(contentTree, entry.getKey(), entry.getValue());
+                    }//
+
+                    content = contentTree.toString();
+                } catch (IOException e) {
+                    log.error("Couldn't read parent JSON {1}", e.getMessage());
+                    // just ignore, we've got gpx data
                     // error coudl be content !, check this
-				}
-    			
+                }
 
             }
 
@@ -276,64 +264,64 @@ public class UploadResource {
         ;
         return view;
     }
-    
+
     private void setTokenValue(JsonNode root, String path, String value) {
-		String[] keys = path.split("/");
-		for(int i = 1; i <keys.length; i++){
-		    String key = keys[i];
-			JsonNode nextToken = root.get(key);
-			if (nextToken == null) {
-				System.out.println(">>> token missing " + key);
-				return;
+        String[] keys = path.split("/");
+        for (int i = 1; i < keys.length; i++) {
+            String key = keys[i];
+            JsonNode nextToken = root.get(key);
+            if (nextToken == null) {
+                System.out.println(">>> token missing " + key);
+                return;
             } else if (nextToken.isValueNode()) {
-			    if (i + 1 == keys.length) {
+                if (i + 1 == keys.length) {
                     break;
                 } else {
                     System.out.println(">>> misplaced token " + key);
-			        return;
+                    return;
                 }
             }
-			root = nextToken;
-		}// for
-        ((ObjectNode) root).put(keys[keys.length-1], value);
-	}
+            root = nextToken;
+        }// for
+        ((ObjectNode) root).put(keys[keys.length - 1], value);
+    }
 
-	private HashMap<String, String> addKeys(String path, JsonNode jsonNode,
-			boolean cloneable) {
-		HashMap<String, String> clonable = new HashMap<String, String>();
+    private HashMap<String, String> addKeys(String path, JsonNode jsonNode,
+            boolean cloneable) {
+        HashMap<String, String> clonable = new HashMap<String, String>();
 
-		if (jsonNode.isObject()) {
-			// recurse
-			ObjectNode objectNode = (ObjectNode) jsonNode;
-			Iterator<Map.Entry<String, JsonNode>> iter = objectNode.fields();
+        if (jsonNode.isObject()) {
+            // recurse
+            ObjectNode objectNode = (ObjectNode) jsonNode;
+            Iterator<Map.Entry<String, JsonNode>> iter = objectNode.fields();
 
-			while (iter.hasNext()) {
-				Map.Entry<String, JsonNode> entry = iter.next();
-				String key = entry.getKey();
-				if (cloneable == true || key.endsWith("_c")) {
-					clonable.putAll(addKeys(path + "/" + entry.getKey(),
-							entry.getValue(), true));
-				} else {
-					clonable.putAll(addKeys(path + "/" + entry.getKey(),
-							entry.getValue(), false));
-				}
-			}
-		} else if (jsonNode.isArray()) {
-			ArrayNode arrayNode = (ArrayNode) jsonNode;
-			/*
-			 * Not Implemented System.out.println("[" + path ); for (int i = 0;
-			 * i < arrayNode.size(); i++) {
-			 * System.out.println(arrayNode.get(i));
-			 * 
-			 * } System.out.println("]" );
-			 */
-		} else if (jsonNode.isValueNode()) {
-			ValueNode valueNode = (ValueNode) jsonNode;
-			if (cloneable) {
-				clonable.put(path, valueNode.asText());
-			}
-		}
+            while (iter.hasNext()) {
+                Map.Entry<String, JsonNode> entry = iter.next();
+                String key = entry.getKey();
+                if (cloneable == true || key.endsWith("_c")) {
+                    clonable.putAll(addKeys(path + "/" + entry.getKey(),
+                            entry.getValue(), true));
+                } else {
+                    clonable.putAll(addKeys(path + "/" + entry.getKey(),
+                            entry.getValue(), false));
+                }
+            }
+        } else if (jsonNode.isArray()) {
+            ArrayNode arrayNode = (ArrayNode) jsonNode;
+            /*
+             * Not Implemented System.out.println("[" + path ); for (int i = 0;
+             * i < arrayNode.size(); i++) {
+             * System.out.println(arrayNode.get(i));
+             *
+             * } System.out.println("]" );
+             */
+        } else if (jsonNode.isValueNode()) {
+            ValueNode valueNode = (ValueNode) jsonNode;
+            if (cloneable) {
+                clonable.put(path, valueNode.asText());
+            }
+        }
 
-		return clonable;
-	}
+        return clonable;
+    }
 }
